requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
sqlalchemy>=1.4.0
schedule>=1.0.0
python-dotenv>=0.19.0
slowapi>=0.1.5
structlog>=21.1.0
prometheus-fastapi-instrumentator>=0.9.0
pydantic>=1.8.0
pydantic-settings